requirements.txt
langchain
langchain-groq
langchain-core
langchain-community
python-dotenv
requests
pydantic
arxiv
kaggle
streamlit
